pyproject.toml
[project]
name = "pytorch-hydra-wandb-template"
version = "0.1.0"
description = ""
authors = [
    {name = "gajitaku-dendai",email = "21aj112@ms.dendai.ac.jp"}
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "wandb (>=0.19.10,<0.20.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "timm (>=1.0.15,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "torchinfo (>=1.8.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "utils", from = "src" },
    { include = "conf", from = "src" },
    { include = "database", from = "src" },
    { include = "architecture", from = "src" },
]
[[tool.poetry.source]]
name = "torch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "torch_cu124"}
torchvision = {source = "torch_cu124"}
torchaudio = {source = "torch_cu124"}
